examples/agent-framework-comparison/requirements.txt
pydantic-ai>=0.0.7
any-agent>=0.1.0
swarm-agent>=0.1.0
langchain>=0.1.0
openai>=1.0.0
anthropic>=0.20.0

# OpenTelemetry
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation>=0.41b0
opentelemetry-exporter-otlp-proto-http>=1.20.0
opentelemetry-semantic-conventions>=0.41b0 